feat: separate prepare/install stages
move versioned pmd-bin-x.y.z folder to unversioned pmd-bin folder to match previous directory structure
