+++ install.js
@@ -13,29 +13,8 @@ async function installJava() {
     });
 }
 
-async function downloadPmd() {
-  const pmdVer = require("./package.json").pmd.version;
-  const url = `https://github.com/pmd/pmd/releases/download/pmd_releases%2F${pmdVer}/pmd-bin-${pmdVer}.zip`;
-  const got = require("got");
-
-  let response = await got(url).buffer();
-
-  return response;
-}
-
-async function extractPmd(buffer) {
-  const Zip = require("adm-zip");
-
-  const pmdZip = new Zip(buffer);
-  pmdZip.extractAllTo("dist", true);
-}
-
-async function installPmd() {
-  return downloadPmd().then(extractPmd);
-}
-
 async function main() {
-  await Promise.all([installJava(), installPmd()]);
+  await installJava();
 }
 
 main();
