+++ pmd
@@ -4,10 +4,9 @@ const path = require("path");
 const jre = require("node-java-connector");
 
 async function main() {
-  const pmdVer = require("./package.json").pmd.version;
   const className = "net.sourceforge.pmd.PMD";
   const classPath = [
-    path.resolve(__dirname, "dist", `pmd-bin-${pmdVer}`, "lib", "*"),
+    path.resolve(__dirname, "dist", "pmd-bin", "lib", "*"),
   ];
   const args = process.argv.splice(2);
